requirements.txt
torch>=2.1
torchvision>=0.16
torchaudio>=2.1 ; platform_system!="Windows"
opencv-python
numpy
pandas
scikit-learn
Pillow
tqdm
